+++ app/src/main/java/org/akvo/caddisfly/Config.java
@@ -46,6 +46,7 @@ public class Config {
     public static final String TRAINING_VIDEO_LINK = "http://caddisfly.ternup.com/akvoapp/caddisfly-training.php";
 
     public static final String CALIBRATE_FOLDER_NAME = "/com.ternup.caddisfly/calibrate/";
+    //public static final String CALIBRATE_FOLDER_NAME = "/com.akvo.caddisfly/saved/";
 
     // Tag for debug log filtering
     public static final String DEBUG_TAG = "Caddisfly";