+++ app/src/main/java/org/akvo/caddisfly/adapter/CalibrateListAdapter.java
@@ -53,15 +53,6 @@ public class CalibrateListAdapter extends ArrayAdapter<ResultRange> {
         @SuppressLint("ViewHolder") View rowView = inflater.inflate(R.layout.row_calibrate, parent, false);
 
         final MainApp mainApp = ((MainApp) activity.getApplicationContext());
-
-//        rowView.setBackgroundResource(position % 2 == 0 ?
-//                R.drawable.listitem_row_2 : R.drawable.listitem_row_1);
-
-//        if (position % 2 == 0) {
-//            rowView.setBackgroundResource(R.drawable.listitem_row_1);
-//        }
-
-
         ArrayList<ResultRange> ranges = mainApp.currentTestInfo.getRanges();
         ResultRange range = ranges.get(position);